app/main/requirements.txt
alembic==1.3.1
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
enum34==1.1.6
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
functools32==3.2.3.post2
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==5.0.0
pathlib2==2.3.5
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
scandir==1.10.0
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
zipp==0.6.0
